Fix bad metadata, properly add CreativeVisRelay
This was an oversight on how parts are generated. The part item damage
is indexed off of the AEPartsEnum file. The CreativeVisInterface was put
in the middle of this enum, so some essentia monitors got converted to
vis relays. This has been reverted so that the CreativeVisInterface is
placed last. Hopefully not too much damage was already done.

Also properly sets the itemstack of the part to the correct one (use the
creative version, not the regular), so that removing the part gives the
creative version back and the part loads as the creative version.

Also adds the creative vis relay interface to the API (because why not).